Refactoring around codegen
